refactored classes, addressed most linter issues
